Update index.md
Do not agree with one of the technical correction requests in this tracker regarding "Home page. 1.2 How to read this Guide. Changes: consider adding a period since the other page titles include periods." None of the titles have period at the end. Reverted this change, so "How to read this Guide" as is with no period.